+++ docs/install.sh
@@ -2,7 +2,7 @@
 set -e
 
 REPO="alexdesousa/planck"
-VERSION="0.1.12"
+VERSION="0.1.13"
 RELEASES="https://github.com/$REPO/releases/latest/download"
 BIN_NAME="planck"