+++ docs/install_docker.sh
@@ -2,7 +2,7 @@
 set -e
 
 REPO="alexdesousa/planck"
-VERSION="0.1.12"
+VERSION="0.1.13"
 PLANCK_HOME="$HOME/planck"
 COMPOSE_URL="https://raw.githubusercontent.com/$REPO/v${VERSION}/planck_docker/compose.yml"